match_multiple_wo_alloc: add match lookahead
Problem: Default match_wo_alloc -n returns sequential
timeslots, like t=89,90,91,etc. This is not as much
information as it could be, because the resources that
are available at t=89 are likely enough to be available
at t=90 that checking is redundant.

Add a match lookahead option to the resource module
that defines the interval over which multiple resources
should be returned.
For example, if lookahead=20 and n=4, the returned resources
could be at t=34, t=39, t=44, and t=49.

Additionally, add a decaying-exponent match_wo_alloc type
that distributes most of the returned timeslots close to the
start time but distributes some farther out.