t4016: Add match-wo-alloc duration extension test
Problem: No test exists for match-without-allocate's ability to extend
the duration of matched resources until their next allocation or the end
of the graph.

Add a testcase to t4014 that compares the expiration of equivalent
match-allocate and match-without-allocating requests to see if and how
match-without-allocating actually extends its match's duration.