t4014: Add match-wo-alloc duration extension test
Problem: No test exists for match-without-allocate's ability to extend
the duration of matched resources until their next allocation or the end
of the graph.

Add a testcase to t4014 that compares the expiration of equivalent
match-allocate and match-without-allocating requests to see if
match-without-allocating actually extends its match's duration.

There is no mechanism to allocate in the future, so we have yet to test
the allocation case.