jobinfo: Add MATCHED state
Problem: For now, match without_allocating consumes a jobid. All
jobids must be queryable, but a match without an allocation is not
stricly a job, nor does it have a state.

Regardless, add a MATCHED state to the jobinfo enum.
This way, the jobid doesn't just disappear to the user, and asking for
information about the jobid returns a sensible result.