trav: Make match_wo_alloc seek first avail match
Problem: A single match-wo-alloc request is not very information-dense
because it is confined to a single start time. In a multicluster
scheduling scheme, such time-constrained match queries might not
provide enough information to be useful.

If a match-without-alloc request fails, try matching in the future
until we find the first available match and return it. In other words,
return the first-available resources on match-wo-alloc requests.