trav: Emit to stringstream in run ()
Problem: To match multiple sets of resources, the traverser needs to be
able to return multiple sets. Because writers do not support recording
multiple sets resources and the emission step occurs outside of the
traverser, this is currently impossible. Rather than make writers
support multiple sets, it is easier to move the emission step into the
traverser. In fact, every time we call tr->run (the matching version)
we already call writers->emit as if it were a part of tr->run.

Make tr->run accept a stringstream and emit to it.
Remove the extra emission step from all calls to tr->run.
When a vector of writers is passed into tr->run, only emit to the
stringstream from the writer at index 0.