match_op: Add string_to_match_op
Problem: In many places in resource/, we switch on a string to create an
appropriate match_op_t. This is prone to error because misspellings are
more likely to occur across so many duplicated strings, and any changes
to the match_op_t enums need to be mirrored correctly across many files.

Create a string_to_match_op function in match_op.h to centralize all
string -> match_op_t mapping.