dfu: Add "within" parameter to match traversals
Problem: Some match options return matches in the future. This is great
because users can get a planned allocation immediately, but matching
resources in the far future is a waste of traversal time if the user
doesn't actually want to wait that long for an allocation.

Add a parameter to traverser->run() that specifies a no-later-than
parameter for returned matches. Match requests with the "within"
parameter will only return matches that start at or before `now` +
`within`.

Add a jobspec user attribute, `match-within`, that specifies this
parameter. Override the user attribute when a value is also specified on
the command line.